it worked

+++ crate/src/color_quad.rs
@@ -14,7 +14,7 @@ pub struct ColorQuad {
 
 impl ColorQuad {
 	pub fn new(render: &Render) -> Self {
-		let vertices: [f64; 12] = [
+		let vertices: [f32; 12] = [
 			-0.5, 0.5, 0.0,
 			-0.5, -0.5, 0.0,
 			0.5, -0.5, 0.0,
@@ -25,7 +25,7 @@ impl ColorQuad {
 		let indices: [u16; 6] = [3, 2, 1, 3, 1, 0];
 		let indices_buf = render.load_index_buffer(&indices);
 
-		let colors: [f64; 12] = [
+		let colors: [f32; 12] = [
 			0.0, 0.0, 0.0,
 			1.0, 0.0, 0.0,
 			0.0, 1.0, 0.0,
